chore: Address review changes
